ref(replay): Refactor loading and error states for Replay Details page
