+++ static/app/views/replays/detail/body/replayDetailsProviders.tsx
@@ -1,0 +1,53 @@
+import {type ReactNode, useEffect} from 'react';
+
+import {LocalStorageReplayPreferences} from 'sentry/components/replays/preferences/replayPreferences';
+import {Provider as ReplayContextProvider} from 'sentry/components/replays/replayContext';
+import useInitialTimeOffsetMs from 'sentry/utils/replays/hooks/useInitialTimeOffsetMs';
+import useLogReplayDataLoaded from 'sentry/utils/replays/hooks/useLogReplayDataLoaded';
+import useMarkReplayViewed from 'sentry/utils/replays/hooks/useMarkReplayViewed';
+import {ReplayPreferencesContextProvider} from 'sentry/utils/replays/playback/providers/replayPreferencesContext';
+import type ReplayReader from 'sentry/utils/replays/replayReader';
+import useOrganization from 'sentry/utils/useOrganization';
+import ReplayTransactionContext from 'sentry/views/replays/detail/trace/replayTransactionContext';
+
+interface Props {
+  children: ReactNode;
+  projectSlug: string | null;
+  replay: ReplayReader;
+}
+
+export default function ReplayDetailsProviders({children, replay, projectSlug}: Props) {
+  const organization = useOrganization();
+
+  const replayRecord = replay.getReplay();
+  const initialTimeOffsetMs = useInitialTimeOffsetMs({
+    orgSlug: organization.slug,
+    projectSlug,
+    replayId: replayRecord.id,
+    replayStartTimestampMs: replayRecord.started_at?.getTime(),
+  });
+
+  const {mutate: markAsViewed} = useMarkReplayViewed();
+  useEffect(() => {
+    if (projectSlug && replayRecord.id && !replayRecord.has_viewed) {
+      markAsViewed({projectSlug, replayId: replayRecord.id});
+    }
+  }, [markAsViewed, organization, projectSlug, replayRecord]);
+
+  useLogReplayDataLoaded({projectId: replayRecord.project_id, replay});
+
+  return (
+    <ReplayPreferencesContextProvider prefsStrategy={LocalStorageReplayPreferences}>
+      <ReplayContextProvider
+        analyticsContext="replay_details"
+        initialTimeOffsetMs={initialTimeOffsetMs}
+        isFetching={false}
+        replay={replay}
+      >
+        <ReplayTransactionContext replayRecord={replayRecord}>
+          {children}
+        </ReplayTransactionContext>
+      </ReplayContextProvider>
+    </ReplayPreferencesContextProvider>
+  );
+}
